implement a per connection rule loader
Rules location can be selected by providing a url with --rules
For each connection, the ruleset will be loaded as json data
and in and out modules will be prepared for this specific connection.

A special case for rules is the fact that the hostname is not
known before we initialize the connection. For this reason,
a "matcher" (a compiled regular expression) is stored in each
module that will be used to know at run time if the module
should be activated or not

For now, only redis URI scheme has been implemented, but file://
and http:// https:// should be straigtforward to implement.