Use SSLContext objects; reuse the SNI
Clients typically fill in the SNI (server name indication) extension to
indicate which (virtual host) they want to connect to, since a server
with one IP address can host several virtual servers. Since the
args.target_ip variable is always normalized to the IP which the given
target resolves to, we use the callback function to retrieve the clients
SNI and reuse that hostname when establishing the connection to the
target.